backend/requirements.txt
mysql-connector-python>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
bcrypt>=4.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
